#10382: Update type 2 complex files
