+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -540,6 +540,31 @@ Complex Operations
 
 The following functions are available,
 
+
+Complex arithmetic can be carried out for multiply, divide, add and subtract as follows:
+
+.. autofunction:: tt_lib.tensor.complex_add
+
+.. autofunction:: tt_lib.tensor.complex_sub
+
+.. autofunction:: tt_lib.tensor.complex_mul
+
+.. autofunction:: tt_lib.tensor.complex_div
+
+and then unary operations for,
+
+.. autofunction:: tt_lib.tensor.real
+
+.. autofunction:: tt_lib.tensor.imag
+
+.. autofunction:: tt_lib.tensor.complex_abs
+
+.. autofunction:: tt_lib.tensor.conj
+
+.. autofunction:: tt_lib.tensor.complex_recip
+
+.. autofunction:: tt_lib.tensor.polar
+
 Complex Operations (Type 2)
 ===========================
 Type 2 Complex representation allows for more flexible storage than earlier one while providing same set of
